+++ template/base_feature/src/main/java/ru/surfstudio/standard/application/logger/strategies/remote/FirebaseCrashlyticsRemoteLoggingStrategy.kt
@@ -1,7 +1,7 @@
-package ru.surfstudio.standard.application.logger
+package ru.surfstudio.standard.application.logger.strategies.remote
 
 import com.google.firebase.crashlytics.FirebaseCrashlytics
-import ru.surfstudio.android.logger.remote_logging_strategies.RemoteLoggingStrategy
+import ru.surfstudio.standard.base.logger.RemoteLoggingStrategy
 
 private const val DEFAULT_STRING_VALUE = "null"