+++ mvi/RELEASE_NOTES.md
@@ -21,3 +21,4 @@ even if stream contains events from UI, which should be reacted directly.
 * Add additional schemes to docs
 * Remove listenForResult extension for EventTransformerList + listenForResult from NavigationMiddleware
 * Fixed StandardReactDialogView negative button nullability
+* ANDDEP-928 Remove Timber dependencies