+++ security/RELEASE_NOTES.md
@@ -8,6 +8,7 @@
     * [ReleaseAppChecker](lib-security/src/main/java/ru/surfstudio/android/security/app/ReleaseAppChecker.kt) - a class with different checks for release application.
     * [ReleaseSignatureChecker](lib-security/src/main/java/ru/surfstudio/android/security/app/ReleaseSignatureChecker.kt) - a class for checking release signature.
     * [SecurityUtils](lib-security/src/main/java/ru/surfstudio/android/security/crypto/security/SecurityUtils.kt) - added different utils for encoding and decoding.
+* ANDDEP-928 Remove clearCrashlyticsLogs function
 ## 0.4.0
 ##### Security
 * ANDDEP-82 Security Module